+++ tests/projects/c++/modules/test_headerunits.lua
@@ -15,26 +15,27 @@ function main(t)
     if is_subhost("windows") then
         local vs = find_vstudio()
         if vs and vs["2022"] then
-            os.exec("xmake f -c")
+            os.exec("xmake f -c --yes")
             _build()
         end
     elseif is_host("linux") then
         local gcc = find_tool("gcc", {version = true})
         if gcc and gcc.version and semver.compare(gcc.version, "11.0") >= 0 then
-            os.exec("xmake f -c")
+            os.exec("xmake f -c --yes")
             _build()
         end
         local clang = find_tool("clang", {version = true})
-        if clang and clang.version and semver.compare(clang.version, "16.0") >= 0 then
-            if is_host("linux") then
+        if clang and clang.version then
+            if semver.compare(clang.version, "15.0") >= 0 then
                 os.exec("xmake clean -a")
                 os.exec("xmake f --toolchain=clang -c")
                 _build()
-            end
+            -- elseif semver.compare(clang.version, "15.0") >= 0 then
             -- there is currently a bug on llvm git that prevent to build STL header units https://github.com/llvm/llvm-project/issues/58540
             -- os.exec("xmake clean -a")
             -- os.exec("xmake f --toolchain=clang --cxxflags=\"-stdlib=libc++\" -c")
             -- _build()
+            end
         end
     end
 end