+++ xmake/actions/install/main.lua
@@ -118,8 +118,11 @@ function main()
                     cprint("${color.success}install ok!")
                     ok = true
                 end
-                if not ok and os.syserror() == os.SYSERR_NOT_PERM then
-                    wprint("please pass the --admin parameter to `xmake install` to request administrator permissions!")
+                if not ok then
+                    local syserror = os.syserror()
+                    if syserror == os.SYSERR_NOT_PERM or syserror == os.SYSERR_NOT_ACCESS then
+                        wprint("please pass the --admin parameter to `xmake install` to request administrator permissions!")
+                    end
                 end
                 assert(ok, "install failed, %s", errors or "unknown reason")
             end