+++ xmake/actions/uninstall/main.lua
@@ -75,8 +75,11 @@ function main()
                     cprint("${color.success}uninstall ok!")
                     ok = true
                 end
-                if not ok and os.syserror() == os.SYSERR_NOT_PERM then
-                    wprint("please pass the --admin parameter to `xmake uninstall` to request administrator permissions!")
+                if not ok then
+                    local syserror = os.syserror()
+                    if syserror == os.SYSERR_NOT_PERM or syserror == os.SYSERR_NOT_ACCESS then
+                        wprint("please pass the --admin parameter to `xmake uninstall` to request administrator permissions!")
+                    end
                 end
                 assert(ok, "uninstall failed, %s", errors or "unknown reason")
             end